+++ src/mruby++/class.h
@@ -16,6 +16,7 @@ public:
     explicit Class(Interpreter* rb, struct RClass* klass);
     ~Class();
 
+    void ivar_set(std::string_view name, std::string_view value);
     void ivar_set(std::string_view name, mrb_value value);
     void ivar_set(mrb_sym name, mrb_value value);