+++ src/rq.rb
@@ -6,6 +6,9 @@ class Rq
   # Returns `true` if the debug flag (`-d`) was enabled.
   def self.debug? = @debug
 
+  # Returns the requested output format (default: `:json`)
+  def self.output_format = @output_format || :json
+
   # Prints the result of `item`.
   def self.print_result(item)
     json_generate_opts = {
@@ -13,13 +16,24 @@ class Rq
       indent_width: 2,
     }
 
-    puts JSON.generate(item, **json_generate_opts)
+    case output_format
+    when :json
+      puts JSON.generate(item, **json_generate_opts)
+    when :ruby
+      # TODO: figure out how to pretty print this
+      puts item.inspect
+    when :plain
+      puts item
+    else
+      raise ArgumentError, "Unknown output format: #{output_format}"
+    end
   end
 
   # Method that returns whatever debug info I like.
   def self.__debug_info = {
     compact: @compact,
     debug: @debug,
+    output_format:,
     ruby: MRUBY_DESCRIPTION,
   }
 end