+++ test/output_format.rb
@@ -1,0 +1,49 @@
+require_relative './support/rq_test'
+
+class OutputFormat < RqTest
+  # BEGIN format tests =======================================================
+  rq_test! :ruby,
+      argv: [
+        '-o',
+        'ruby',
+      ],
+      stdin:  :basic_array,
+      stdout: :from_fixture
+  # END format tests =========================================================
+
+  # BEGIN flag handling ======================================================
+  rq_test! :no_arg,
+      argv: [
+        '-o',
+      ],
+      status: 1,
+      stderr: :from_fixture,
+      msg: '-o needs an argument'
+
+  rq_test! :combined_flag_at_start,
+      argv: [
+        '-od',
+        'ruby'
+      ],
+      status: 1,
+      stderr: :from_fixture,
+      msg: '-o needs to be the last one'
+
+  rq_test! :combined_flag_at_end,
+      argv: [
+        '-do',
+        'ruby',
+      ],
+      stdin:  :basic_array,
+      stdout: :test_output_format_ruby,
+      stderr: :from_fixture
+
+  rq_test! :combined_arg,
+      argv: [
+        '-oruby',
+      ],
+      status: 1,
+      stderr: :test_output_format_combined_flag_at_start,
+      msg: 'format parameter needs to be its own argument'
+  # END flag handling ========================================================
+end