+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## main
 
 - Add `-c` option for emitting compact JSON
+- Add `-o` option for specifying the output format, e.g. `-o ruby` or
+  `-o plain` (defaults to `-o json`)
 
 ## v0.1.0